Exception handling, php installer, readme changes

+++ scripts/remove-vhost.sh
@@ -1,5 +1,3 @@
-#!/bin/bash
-
 FLAG=true
 PROJECT_NAME=''
 
@@ -7,16 +5,31 @@ while $FLAG
 do                
         read -p "Enter name of virtual host you want to remove (e.g. myproject): " PROJECT_NAME
 
-        if [  -f "/etc/apache2/sites-available/$PROJECT_NAME.local.conf" ]; then
-                FLAG=false
-                rm /etc/apache2/sites-available/$PROJECT_NAME.local.conf
-                service apache2 reload
-                echo "Virtual host has beed removed."
+        # Validate input
+        VALIDATED="$(echo $PROJECT_NAME | grep '^[a-z0-9]\+\-\?[a-z0-9]\+$')"
+
+        # Check if input is correct
+        if [ "$PROJECT_NAME" = "$VALIDATED" ] ; then
+
+                # Check if given vhost exists
+                if [  -f "/etc/apache2/sites-available/$PROJECT_NAME.local.conf" ]; then
+                        
+                        # Remove vhost and reload Apache
+                        rm /etc/apache2/sites-available/$PROJECT_NAME.local.conf
+                        service apache2 reload
+                        
+                        echo "Virtual host has beed removed."
+
+                        # Set flag to false to leave loop
+                        FLAG=false
+                else
+                        echo "${RED}Virtual host does not exist.${CLC} Check if virtual host name is correct."
+                fi
+
         else
-                FLAG=true
-                echo "${RED}Virtual host already exists.${CLC} Choose different virtual host name."
+                echo "${RED}Incorrect input.${CLC} Try again."
         fi
 done
 
-#Sleep to let user read console
+# Sleep to let user read console
 sleep 4
